Add RabbitMQ user/vhost configuration
By default, we just use the guest user. But, this makes it possible to
add or remove vhosts and users from rabbitmq.

This requires fairly explicit rabbitmq configuration options.

Originally, this used a map filter to extract vhosts from the list of
users and check for guest in the list of users, but the map filter is
not available in EL6 with Jinja2 2.6, so we can't rely on using that
here. Instead of extracting vhosts, or checking the list of users to
determine whether or not to remove guest, we require explicit
configuration of users and vhosts.