Make sure mistral db gets upgraded too
When upgrading mistral, the database needs to be updated as well.
This adds the version number to the signal file so that on upgrade,
it will run the 'upgrade head' and 'populate' commands again.

Note that the new 'Enable mistral db upgrade' task skips linting,
because it must be done at that point not when handlers are flushed.
So, turning that task into a handler doesn't make sense.