Ansible 2.5 | Use ansible_facts namespaced vars
With 2.5, facts are now namespaced under ansible_facts.[1]
They are still accessible under the ansible_* vars, however that can be
disabled by setting `inject_facts_as_vars` to False. The default is True
in 2.5, but "is expected to be set to 'False' in a future release.
Migrating now means these roles and playbooks do not depend on this
setting being True.

[1] https://docs.ansible.com/ansible/2.7/porting_guides/porting_guide_2.5.html#ansible-fact-namespacing