Use yum_repository to install nodesource repo
The nodesource repo rpm causes problems if a node.js major version
upgrade is required (such as from 4.x to 10.x) because the version
number of the repo version does not change even though the node.js
version in the repo file changes.

Rather than hacking around that problematic rpm, we use the
yum_repository ansible module to install the repo.